number of simulated objects 870 -> 90
This stops the game from chugging really hard when it first starts.

From brief inspection in the physx visual debugger, it appears that the
chugging happened because physx was simulating collisions for so many
things at once, and the speedup would happen as each of the colliders
was put to sleep.