Added an imgui window to switch camera modes
The imgui panel is in graphic system.cpp
Clicking the buttons switches an int value.

This int value is used in the update function and depending on the int value it uses different V values which correspond to different camera modes
(Free, fixed) - can be extended to toggle follow camera.

I had to put the cam_mode int into the GraphicSystem header file, and initialize the V glm::mat4 in the header file for this to work.

I then call gs.ImGuiPanel in main for the panel to show up and the changes when the buttons are pressed to propagate properly