Changed jump condition & move follow camera back
The jump was not triggering on certain parts of the level because it was checking for the vehicle being below a certain y value - the new level went above that value, so I changed the condition to check if the car is on the ground. - future proofs it - this is what I originally wanted to do months ago but didn't know how to at the time.

I also moved the follow camera back so you see a bit more of the level and skybox, and in doing so makes the car feel faster because the FOV is larger