Changed AI jump by adding new function
The ai was jumping too high using the player function for some reason. I gave the AI a new function with reduced forces.

There was also a function declaration in Car.h that had Car:: in front of it.

for the setup1() so the ai was not being given the proper driving model